Update update_user to use sp_updateuser stored procedure with correct parameter order
