+++ .env
@@ -5,4 +5,10 @@
 # Example for SQL Server:
 # DATABASE_URL=mssql+pyodbc://username:password@server_name/dbname?driver=ODBC+Driver+17+for+SQL+Server
 
+AZURE_SQL_SERVER="localhost"
+AZURE_SQL_DATABASE="SMS"
+AZURE_SQL_USERNAME="sa"
+AZURE_SQL_PASSWORD="Alex@123#"
+
+
 DATABASE_URL=postgresql://postgres:password@localhost:5432/user_management
