+++ app/utils/db_connection.py
@@ -1,0 +1,17 @@
+import os
+import pyodbc
+from fastapi import HTTPException
+
+def get_connection():
+    server = os.getenv("AZURE_SQL_SERVER")
+    database = os.getenv("AZURE_SQL_DATABASE")
+    username = os.getenv("AZURE_SQL_USERNAME")
+    password = os.getenv("AZURE_SQL_PASSWORD")
+    driver = '{ODBC Driver 18 for SQL Server}'
+
+    connection_string = f'DRIVER={driver};SERVER={server};DATABASE={database};UID={username};PWD={password};Encrypt=yes;TrustServerCertificate=yes;'
+    try:
+        conn = pyodbc.connect(connection_string)
+        return conn
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Database connection error: {str(e)}")
